Working on more elegant SVG solution
Basically, I am trying to find a solution for detecting field presses
that requires as little 'switch-case' with 83 cases as possible, so I
wanted to encode everything into the SVG, so the logic of server and
backend can be reduced. Note that some things are not working right now,
I'll fix them before I merge this into the master branch.